Update exec tests (#185)
* Corrected string encoding in the execution test suite harness.

* Updated failing execution tests: expressions directory.

* Updated failing execution tests: variables directory.

* Updated failing execution tests: operators directory.

* Updated failing execution tests: array/push directory.

* Updated failing execution tests: array/indexAccess directory.

* Updated failing execution tests: array/pop directory.

* Updated failing execution tests: array/delete directory.

* Updated failing execution tests: array/copying directory.

* Updated failing execution tests: array directory.

+++ test/libsolidity/semanticTests/array/copying/array_of_structs_containing_arrays_memory_to_storage.sol
@@ -9,7 +9,7 @@ contract C {
 
     S[] s;
 
-    function f() external returns (uint256, uint256, uint128, uint128) {
+    function f() external returns (uint256, uint, uint128, uint128) {
         S[] memory m = new S[](3);
         m[1] = S(0, [uint128(1), 2, 3], new uint128[](3));
         m[1].c[0] = 1;
