+++ test/libsolidity/semanticTests/array/function_memory_array.sol
@@ -1,25 +1,25 @@
 contract C {
-    function a(uint256 x) public returns (uint256) {
+    function a(uint x) public returns (uint) {
         return x + 1;
     }
 
-    function b(uint256 x) public returns (uint256) {
+    function b(uint x) public returns (uint) {
         return x + 2;
     }
 
-    function c(uint256 x) public returns (uint256) {
+    function c(uint x) public returns (uint) {
         return x + 3;
     }
 
-    function d(uint256 x) public returns (uint256) {
+    function d(uint x) public returns (uint) {
         return x + 5;
     }
 
-    function e(uint256 x) public returns (uint256) {
+    function e(uint x) public returns (uint) {
         return x + 8;
     }
 
-    function test(uint256 x, uint256 i) public returns (uint256) {
+    function test(uint x, uint i) public returns (uint) {
         function(uint) internal returns (uint)[] memory arr =
             new function(uint) internal returns (uint)[](10);
         arr[0] = a;
@@ -34,9 +34,9 @@ contract C {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// test(uint256,uint256): 10, 0 -> 11
-// test(uint256,uint256): 10, 1 -> 12
-// test(uint256,uint256): 10, 2 -> 13
-// test(uint256,uint256): 10, 3 -> 15
-// test(uint256,uint256): 10, 4 -> 18
-// test(uint256,uint256): 10, 5 -> FAILURE, hex"4e487b71", 0x51
+// test(uint,uint): 10, 0 -> 11
+// test(uint,uint): 10, 1 -> 12
+// test(uint,uint): 10, 2 -> 13
+// test(uint,uint): 10, 3 -> 15
+// test(uint,uint): 10, 4 -> 18
+// test(uint,uint): 10, 5 -> FAILURE, 1
