+++ test/libsolidity/semanticTests/array/pop/byte_array_pop.sol
@@ -1,7 +1,7 @@
 contract c {
     bytes data;
 
-    function test() public returns (uint256 x, uint256 y, uint256 l) {
+    function test() public returns (uint x, uint y, uint l) {
         data.push(0x07);
         data.push(0x03);
         x = data.length;
